% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_datasets.R
\name{cds_datasets}
\alias{cds_datasets}
\title{CDS dataset list}
\usage{
cds_datasets(user, simplify = TRUE, verbose = FALSE)
}
\arguments{
\item{user}{string, user ID used to sign up for the CDS data service and
used to retrieve the token set by \code{\link[ecmwfr]{cds_set_key}}.
Note: can also be set to \code{NULL}, in this case user and key will
be read from the \code{.cdsapirc} file (located in your home folder).}

\item{simplify}{simplify the output, logical (default = \code{TRUE})}

\item{verbose}{boolean, default \code{FALSE}.}
}
\value{
returns a nested list or data frame with the ECMWF datasets
}
\description{
Returns a list of datasets available on the climate data store servers.
}
\examples{

\donttest{
# get a list of services
cds_services(user = NULL)
}
}
\seealso{
\code{\link[ecmwfr]{cds_set_key}}
\code{\link[ecmwfr]{cds_transfer}}
\code{\link[ecmwfr]{cds_request}}
}
\author{
Reto Stauffer
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
