% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_delete.R
\name{cds_delete}
\alias{cds_delete}
\title{CDS delete request}
\usage{
cds_delete(user, url, verbose = TRUE)
}
\arguments{
\item{user}{character, user ID to sign up for the CDS data service and
used to retrieve the token set by \code{\link[ecmwfr]{cds_set_key}}.
Can also be \code{NULL} if you preferr to use the \code{.cdsapirc} file.}

\item{url}{task url for request}

\item{verbose}{show feedback on processing}
}
\description{
Deletes a staged download from the queue
}
\seealso{
\code{\link[ecmwfr]{cds_set_key}}
\code{\link[ecmwfr]{cds_transfer}}
\code{\link[ecmwfr]{cds_request}}
}
\author{
Reto Stauffer
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
