% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_key.R
\name{cds_get_key}
\alias{cds_get_key}
\title{Get secret CDS token}
\usage{
cds_get_key(user)
}
\arguments{
\item{user}{chracter, user id used to sign up for the CDS data server}
}
\value{
the key set using \code{\link[ecmwfr]{cds_set_key}} saved
in the keychain
}
\description{
Returns you token set by \code{\link[ecmwfr]{cds_set_key}}.
}
\details{
Wrapper function for \code{\link[ecmwfr]{wf_set_key}}.
}
\examples{

\donttest{
# set key
cds_set_key(user = "1234", key = "abc123foo")

# get key
cds_get_key(user = "1234")
}

}
\seealso{
\code{\link[ecmwfr]{cds_set_key}}, \code{\link[ecmwfr]{cds_key_from_file}}
}
\author{
Koen Kufkens, Reto Stauffer
}
\keyword{key}
\keyword{management}
