% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_key.R
\name{cds_key_from_file}
\alias{cds_key_from_file}
\title{Get secret CDS token from file}
\usage{
cds_key_from_file(verbose = FALSE)
}
\arguments{
\item{verbose}{boolean, default is \code{FALSE}.}
}
\value{
Returns a list with the \code{user}, \code{email}, and \code{key}.
 \code{email == user}, used to simplify the calls.
}
\description{
Reading username and secret key from .cdsapirc file
(located in user home directory).
}
\examples{
\donttest{
cds_key_from_file()
}

}
\seealso{
\code{\link[ecmwfr]{cds_set_key}}, \code{\link[ecmwfr]{cds_get_key}}
}
\author{
Reto Stauffer
}
\keyword{key}
\keyword{management}
