% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_request.R
\name{cds_request}
\alias{cds_request}
\title{CDS data request and download}
\usage{
cds_request(user, request, transfer = TRUE, path = tempdir(),
  time_out = 3600, verbose = TRUE)
}
\arguments{
\item{user}{username you used to register the API key
when calling \code{\link[ecmwfr]{cds_set_key}}. Note: can also be set to \code{NULL}
to use the \code{.cdsapirc} located in your home directory (if existing).}

\item{request}{nested list with query parameters following the layout
as specified on the CDS API page}

\item{transfer}{logical, download data \code{TRUE} or \code{FALSE}
(default \code{transfer = FALSE})}

\item{path}{path were to store the downloaded data}

\item{time_out}{how long to wait on a download to start (default = 3600)}

\item{verbose}{show feedback on processing}
}
\value{
a download query staging url or a netCDF of data on disk
}
\description{
Stage a data request, and optionally download the data to disk. Alternatively
you can only stage requests, logging the request URLs to submit download
queries later on using \code{\link[ecmwfr]{wf_transfer}}. The function only
allows NetCDF downloads, and will override calls for grib data.
}
\examples{

\donttest{
# Use local keyring:
cds_set_key(user = "1234", key = "abcd1345foo")

# Specify request 
era_request <- list(
            "dataset" = "reanalysis-era5-pressure-levels",
            "product_type" = "reanalysis",
            "format" = "netcdf",
            "variable" = "temperature",
            "pressure_level" = "850",
            "year" = "2000",
            "month" = "04",
            "day" = "04",
            "time" = "00:00",
            "area" = "70/-20/00/60",
            "format" = "netcdf",
            "target" = "era5-demo.nc")

# Request data/download file using 'user = "1234"'.
# Note: requires that you have stored the login information
# via 'cds_set_key()' (see above).
cds_request(user = "1234",           # user ID (for authentification)
            request = era_request,   # the request
            transfer = TRUE,         # download the file
            path = ".")              # store data in current working directory

# (Ugly) demo plot to see if we got what we expect,
# requires the ncdf4 library to be installed.
nc <- ncdf4::nc_open("era5-demo.nc")
image(sort(ncdf4::ncvar_get(nc, "longitude")),
      sort(ncdf4::ncvar_get(nc, "latitude")),
      ncdf4::ncvar_get(nc, "t"))
ncdf4:nc_close(nc)

# Alternatively (as an alternative to use local keyring):
# store user information in .cdsapirc and set user = NULL
# to use login information from .cdsapirc:
cds_request(user = NULL,   # use .cdsapirc
            request = era_request,   # the request
            transfer = TRUE,         # download the file
            path = ".")              # store data in current working directory

# (Ugly) demo plot to see if we got what we expect,
# requires the ncdf4 library to be installed.
nc <- ncdf4::nc_open("era5-demo.nc")
image(sort(ncdf4::ncvar_get(nc, "longitude")),
      sort(ncdf4::ncvar_get(nc, "latitude")),
      ncdf4::ncvar_get(nc, "t"))
ncdf4:nc_close(nc)
}
}
\seealso{
\code{\link[ecmwfr]{cds_set_key}},
\code{\link[ecmwfr]{cds_get_key}},
\code{\link[ecmwfr]{cds_key_from_file}},
\code{\link[ecmwfr]{wf_transfer}}
}
\author{
Koen Kufkens, Reto Stauffer
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
