% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_key.R
\name{cds_set_key}
\alias{cds_set_key}
\title{Set secret CDS token}
\usage{
cds_set_key(user, key)
}
\arguments{
\item{user}{chracter, user id used to sign up for the CDS data server}

\item{key}{token provided by CDS}
}
\description{
Saves the token to your local keychain under
a service called "ecmwfr".
}
\details{
Wrapper function for \code{\link[ecmwfr]{wf_set_key}}.
}
\examples{

\donttest{
# set key
cds_set_key(user = "1234", key = "abc123foo")

# get key
cds_get_key(user = "1234")
}
}
\seealso{
\code{\link[ecmwfr]{cds_get_key}}, \code{\link[ecmwfr]{cds_key_from_file}}
}
\author{
Koen Kufkens, Reto Stauffer
}
\keyword{key}
\keyword{management}
