% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_request.R
\name{check_request}
\alias{check_request}
\title{Check requests}
\usage{
check_request(x)
}
\arguments{
\item{x}{list of parameters (as for \code{\link[ecmwfr]{wf_request}})}
}
\value{
Returns the request (might, to avoid problems, modify the
user request.
}
\description{
There are some parameter combinations which are not allowed
and will result in an error when sending the requests to
via API. the \code{\link[ecmwfr]{check_request}} function tries
to avoid these problems before sending a request. See 'Details'.
}
\details{
Some of the checks:
\itemize{
   \item will raise an error if the \code{dataset} identifier is missing.
   \item if \code{dataset = "mars"} and \code{format = "netcdf"}: if no
      \code{grid} specification is given: stop. Reason: grib to netcdf
      conversion will only work on a regular_ll grid.
}
}
\author{
Reto Stauffer
}
