% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwf_download.R
\name{ecmwf_download}
\alias{ecmwf_download}
\title{ECMWF download request}
\usage{
ecmwf_download(email, url, path = tempdir(), filename = "ecmwf_tmp.nc")
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{ecmwf_set_key}}}

\item{url}{url to query}

\item{path}{path were to store the downloaded data}

\item{filename}{filename to use for the downloaded data}
}
\value{
a netCDF of data on disk as specified by a
\code{\link[ecmwfr]{ecmwf_request}}
}
\description{
Returns the contents of the requested url as a netCDF file downloaded
to disk.
}
\examples{

\donttest{
# set key
ecmwf_set_key(email = "test@mail.com", key = "123")

# get key
ecmwf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{ecmwf_set_key}}
\code{\link[ecmwfr]{ecmwf_download}}
\code{\link[ecmwfr]{ecmwf_request}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
