% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwf_get_key.R
\name{ecmwf_get_key}
\alias{ecmwf_get_key}
\title{Get secret ECMWF token}
\usage{
ecmwf_get_key(email, key)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service}
}
\value{
the key set using ecmwf_set_key() saved in the keychain
}
\description{
Returns you token set by ecmwf_set_key()
}
\examples{

\donttest{
# set key
ecmwf_set_key(email = "test@mail.com", key = "123")

# get key
ecmwf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{ecmwf_set_key}}
}
\keyword{key}
\keyword{management}
