% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwf_request.R
\name{ecmwf_request}
\alias{ecmwf_request}
\title{ECMWF data request and download}
\usage{
ecmwf_request(email, path = tempdir(), time_out = 3600,
  download = FALSE, request = list(stream = "oper", levtype = "sfc",
  param = "165.128/166.128/167.128", dataset = "interim", step = "0", grid
  = "0.75/0.75", time = "00/06/12/18", date = "2014-07-01/to/2014-07-31",
  type = "an", class = "ei", area = "73.5/-27/33/45", format = "netcdf",
  target = "tmp.nc"))
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{ecmwf_set_key}}}

\item{path}{path were to store the downloaded data}

\item{time_out}{how long to wait on a download to start}

\item{download}{logical, download data TRUE or FALSE (default = FALSE)}

\item{request}{nested list with query parameters following the layout
as specified on the ECMWF API page}
}
\value{
a download query staging url or a netCDF of data on disk
}
\description{
Stage a data request, and optionally download the data to disk. Alternatively
you can only stage requests, logging the request URLs to submit download
queries later on using \code{\link[ecmwfr]{ecmwf_download}}
}
\examples{

\donttest{
# set key
ecmwf_set_key(email = "test@mail.com", key = "123")

# get key
ecmwf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{ecmwf_set_key}}
\code{\link[ecmwfr]{ecmwf_download}}
\code{\link[ecmwfr]{ecmwf_status}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
