% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ecmwf_server}
\alias{ecmwf_server}
\alias{cds_server}
\title{load default server names}
\usage{
ecmwf_server(id)

cds_server(id)
}
\description{
load default server names
}
\author{
Koen Kufkens

Reto Stauffer
}
