% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmwf_status.R
\name{ecmwf_status}
\alias{ecmwf_status}
\title{ECMWF status request}
\usage{
ecmwf_status(email, url)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{ecmwf_set_key}}}

\item{url}{url to query}
}
\value{
returns a nested list of download information
}
\description{
Returns the contents of a request url, useful when checking staged
downloads which were not directly downloaded by ecmwf_download.
}
\examples{

\donttest{
# set key
ecmwf_set_key(email = "test@mail.com", key = "123")

# get key
ecmwf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{ecmwf_set_key}}
\code{\link[ecmwfr]{ecmwf_download}}
\code{\link[ecmwfr]{ecmwf_request}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
