% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{spinner}
\alias{spinner}
\title{Simple progress spinner}
\usage{
spinner(seconds, id)
}
\arguments{
\item{seconds}{integer, seconds to sleep}

\item{id}{character missing (default) or a string which will
be displayed as 'id'.}
}
\description{
Shows a spinner while waiting for a request to be processed.
}
\author{
Koen Kufkens, Reto Stauffer
}
