% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_check_request.R
\name{wf_check_request}
\alias{wf_check_request}
\title{check ECMWF / CDS data requests}
\usage{
wf_check_request(request)
}
\arguments{
\item{request}{nested list with query parameters following the layout
as specified on the ECMWF API page}
}
\value{
a data frame with the determined service and url service endpoint
}
\description{
Check the validity of a data request by comparing the main dataset
to the list provided by \code{\link[ecmwfr]{wf_datasets}}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_transfer}},\code{\link[ecmwfr]{wf_request}},
\code{\link[ecmwfr]{wf_transfer}}
}
\author{
Koen Hufkens
}
