% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_key_from_file.R
\name{wf_key_from_file}
\alias{wf_key_from_file}
\title{Get secret ECMWF token from file}
\usage{
wf_key_from_file(verbose = FALSE)
}
\arguments{
\item{verbose}{boolean, default \code{FALSE}.}
}
\value{
Returns a list with the \code{email} name and \code{key} from
the .ecmwfapirc file.
}
\description{
Reading email and secret key from .ecmwfapirc file
(located in user home directory).
}
\examples{
\donttest{
ef_key_from_file()
}

}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}, \code{\link[ecmwfr]{wf_get_key}}
}
\author{
Reto Stauffer
}
\keyword{key}
\keyword{management}
