% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_modify_request.R
\name{wf_modify_request}
\alias{wf_modify_request}
\title{Dynamically modify MARS / CDS request parameters}
\usage{
wf_modify_request(request, ...)
}
\arguments{
\item{request}{a MARS or CDS request formatted as an R list()}

\item{...}{variables to replace in the provided rqquest}
}
\value{
a MARS / CDS request
}
\description{
Provides a way to quickly modify MARS / CDS request parameters. The function
will only replace existing parameters, all other parameters specified will
be ignored.
}
\examples{
\dontrun{
# a MARS request for webapi ECMWF data
base_request <- list(stream = "oper",
levtype = "sfc",
param = "165.128/166.128/167.128",
dataset = "interim",
step = "0",
grid = "0.75/0.75",
time = "00/06/12/18",
date = "2014-07-01/to/2014-07-31",
type = "an",
class = "ei",
area = "73.5/-27/33/45",
format = "netcdf",
target = "tmp.nc")


# add an additional month of data and grow the extent by a degree
# from 45 to 46 degrees North
new_request <- wf_modify_request(request = base_request,
                                date = "2014-07-01/to/2014-08-31",
                                area = "73.5/-27/33/46")
print(str(new_request))
}
}
