% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_services.R
\name{wf_services}
\alias{wf_services}
\title{ECMWF services list}
\usage{
wf_services(email)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}}

\item{simplify}{simplify the output, logical (default = TRUE)}
}
\value{
returns a nested list or data frame with the ECMWF services
}
\description{
Returns a list of services
}
\examples{

\donttest{
# set key
wf_set_key(email = "test@mail.com", key = "123")

# get key
wf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_transfer}}
\code{\link[ecmwfr]{wf_request}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
