% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_status.R
\name{wf_status}
\alias{wf_status}
\title{ECMWF status request}
\usage{
wf_status(email, url)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}}

\item{url}{url to query}
}
\value{
returns a nested list of download information
}
\description{
Returns the contents of a request url, useful when checking staged
downloads which were not directly downloaded by ecmwf_download.
}
\examples{

\donttest{
# set key
wf_set_key(email = "test@mail.com", key = "123")

# get key
wf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_transfer}}
\code{\link[ecmwfr]{wf_request}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
