% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_transfer.R
\name{wf_transfer}
\alias{wf_transfer}
\title{ECMWF data transfer function}
\usage{
wf_transfer(email, url, service = "webapi", path = tempdir(),
  filename = basename(tempfile("ecmwfr_", fileext = ".nc")),
  verbose = TRUE)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}} when
calling \code{\link[ecmwfr]{wf_transfer}}.
If set to \code{NULL} the \code{.ecmwfapirc} file will be used.}

\item{url}{url to query}

\item{service}{which service to use, one of \code{webapi} or \code{cds}}

\item{path}{path were to store the downloaded data}

\item{filename}{filename to use for the downloaded data}

\item{verbose}{show feedback on data transfers}

\item{user}{string, user ID when calling \code{\link[ecmwfr]{cds_transfer}}.
If set to \code{NULL} the \code{.cdsapirc} file will be used.}

\item{...}{forwarded to \code{\link[ecmwfr]{wf_transfer}}}
}
\value{
a netCDF of data on disk as specified by a
\code{\link[ecmwfr]{wf_request}}
}
\description{
Returns the contents of the requested url as a netCDF file downloaded
to disk or the current status of the requested transfer.
}
\examples{

\dontrun{
# set key
wf_set_key(email = "test@mail.com", key = "123")

# request data and grab url and try a transfer
r <- wf_request("test@email.com")
wf_transfer("test@email.com", url = r$href)
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_request}}
}
\author{
Koen Kufkens
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
