% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_productinfo.R
\name{cds_productinfo}
\alias{cds_productinfo}
\title{CDS product list for a given dataset}
\usage{
cds_productinfo(user, dataset, show = TRUE, verbose = FALSE)
}
\arguments{
\item{user}{string, user ID used to sign up for the CDS data service and
used to retrieve the token set by \code{\link[ecmwfr]{cds_set_key}}.
Note: can also be set to \code{NULL}, in this case user and key will
be read from the \code{.cdsapirc} file (located in your home folder).}

\item{dataset}{character, name of the data set for which the product
information should be loaded.}

\item{show}{boolean, default \code{TRUE}. If \code{TRUE} the description
will be shown in your browser.}

\item{verbose}{boolean, default \code{FALSE}.}
}
\value{
Downloads a (html) product description from CDS. If
\code{show = FALSE} a list with product details will be returned.
If \code{show = TRUE} the product information will be shown in a local
browser window (plus invisible return of the details as if \code{show}
whould be set \code{FALSE}).
}
\description{
Shows and returns detailed product information about a specific data set
(see \code{\link[ecmwfr]{cds_datasets}}).
}
\examples{
\donttest{
   # Opend description in browser
   cds_productinfo(NULL, "reanalysis-era5-single-levels")

   # Return information
   info <- cds_productinfo(NULL, "reanalysis-era5-single-levels", show = FALSE)
   names(info)
}
}
\seealso{
\code{\link[ecmwfr]{cds_datasets}}.
}
\author{
Reto Stauffer
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
