% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ecmwfr_archetype}
\alias{print.ecmwfr_archetype}
\title{Methods to deal with visualizing / printing
requesting info from the archetype constructor}
\usage{
\method{print}{ecmwfr_archetype}(x, ...)
}
\arguments{
\item{x}{archetype object}

\item{...}{additional parameters to pass on}
}
\description{
Methods to deal with visualizing / printing
requesting info from the archetype constructor
}
