% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_archetype.R
\name{wf_archetype}
\alias{wf_archetype}
\title{Creates an archetype function}
\usage{
wf_archetype(request, ...)
}
\arguments{
\item{request}{a MARS or CDS request as an R list object, with certain
parameters modified to take arguments.}

\item{...}{list of arguments for which a default value can be set}
}
\value{
a MARS / CDS request
}
\description{
Creates a universal MARS / CDS formatting function, in ways
similar to \code{wf_modify_request()} but the added advantage
that you could code for the use of dynamic changes in the
parameters provided to the resulting custom function.
}
\details{
Contrary to a simple replacement as in \code{wf_modify_request()} the
generated functions are considered custom user written. Given the potential
for complex formulations and formatting commands NO SUPPORT for the
resulting functions can be provided. Only the generation of a valid function
will be guaranteed and tested for.
}
\examples{
\dontrun{
# format an archetype function
ERA_interim <- wf_archetype(
 list(class = "ei",
    dataset = "interim",
    expver = "1",
    levtype = "pl",
    stream = "moda",
    type = "an",
    format = "netcdf",
    date = date,
    grid = paste0(res, "/", res),
    levelist = levs,
    param = "155.128",
    target = "output"),
res = 3  # sets default argument
)

# print output of the function with below parameters
str(ERA_interim("20100101", 3, 200))

}
}
