% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_get_key.R
\name{wf_get_key}
\alias{wf_get_key}
\title{Get secret ECMWF / CDS token}
\usage{
wf_get_key(user = "ecmwfr")
}
\arguments{
\item{user}{user (email address) used to sign up for the ECMWF data service}
}
\value{
the key set using \code{\link[ecmwfr]{wf_set_key}} saved
in the keychain
}
\description{
Returns you token set by \code{\link[ecmwfr]{wf_set_key}}
}
\examples{

\dontrun{
# set key
wf_set_key(key = "123")

# get key
wf_get_key()
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
}
\author{
Koen Hufkens
}
