% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_get_key.R
\name{wf_get_key}
\alias{wf_get_key}
\title{Get secret ECMWF / CDS token}
\usage{
wf_get_key(user, service = "webapi")
}
\arguments{
\item{user}{user (email address) used to sign up for the ECMWF data service}

\item{service}{which service to use, one of \code{webapi}, \code{cds}
or \code{ads} (default = webapi)}
}
\value{
the key set using \code{\link[ecmwfr]{wf_set_key}} saved
in the keychain
}
\description{
Returns you token set by \code{\link[ecmwfr]{wf_set_key}}
}
\examples{

\dontrun{
# set key
wf_set_key(user = "test@mail.com", key = "123")

# get key
wf_get_key(user = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
}
\author{
Koen Kufkens
}
