% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_set_key.R
\name{wf_set_key}
\alias{wf_set_key}
\title{Set secret ECMWF token}
\usage{
wf_set_key(email, key, service = "webapi")
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service}

\item{key}{token provided by ECMWF}

\item{service}{which service to use}
}
\description{
Saves the token to your local keychain under
a service called "ecmwfr".
}
\examples{

\dontrun{
# set key
wf_set_key(email = "test@mail.com", key = "123")

# get key
wf_get_key(email = "test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_get_key}}
}
\author{
Koen Kufkens
}
\keyword{key}
\keyword{management}
