% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_user_info.R
\name{wf_user_info}
\alias{wf_user_info}
\title{ECMWF WebAPI user info query}
\usage{
wf_user_info(email)
}
\arguments{
\item{email}{email address used to sign up for the ECMWF data service and
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}}
}
\value{
returns a data frame with user info
}
\description{
Returns user info for the ECMWF WebAPI
}
\examples{

\dontrun{
# set key
wf_set_key(email = "test@mail.com", key = "123")

# get user info
wf_user_info("test@mail.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_services}}
\code{\link[ecmwfr]{wf_datasets}}
}
\keyword{climate,}
\keyword{data}
\keyword{download,}
\keyword{re-analysis}
